bot/requirements.txt
botbuilder-integration-aiohttp>=4.14.0
requests>=2.31.0
botbuilder-dialogs>=4.15.0
aiohttp
botbuilder-core
botbuilder-dialogs
botbuilder-schema
azure-cosmos
